Run observable tests in paralell
